+++ s/wait/parts/get.ts
@@ -1,16 +1,16 @@
 
-import {getErr, getOk, needErr, needOk} from "@e280/stz"
+import {getErr, getOk, gotOk, gotErr} from "@e280/stz"
 import {Wait} from "./type.js"
 
-export function waitGetOk<Value>(wait: Wait<Value>) {
+export function waitGet<Value>(wait: Wait<Value>) {
 	return wait.done
 		? getOk(wait)
 		: undefined
 }
 
-export function waitNeedOk<Value>(wait: Wait<Value>) {
+export function waitGot<Value>(wait: Wait<Value>) {
 	if (!wait.done) throw new Error("wait not done")
-	return needOk(wait)
+	return gotOk(wait)
 }
 
 export function waitGetErr<E = unknown>(wait: Wait<unknown, E>) {
@@ -19,8 +19,17 @@ export function waitGetErr<E = unknown>(wait: Wait<unknown, E>) {
 		: undefined
 }
 
-export function waitNeedErr<E = unknown>(wait: Wait<unknown, E>) {
+export function waitGotErr<E = unknown>(wait: Wait<unknown, E>) {
 	if (!wait.done) throw new Error("wait not done")
-	return needErr(wait)
+	return gotErr(wait)
 }
+
+/** @deprecated renamed to `waitGet` */
+export const waitGetOk = waitGet
+
+/** @deprecated renamed to `waitGot` */
+export const waitNeedOk = waitGot
+
+/** @deprecated renamed to `waitGotErr` */
+export const waitNeedErr = waitGotErr